GW-25: Refactored folder structure
