+++ src/pages/home/Home.js
@@ -2,8 +2,8 @@ import React from "react";
 import About from "./components/About";
 import { MainContainer } from "../../shared/layout/MainContainer";
 import { LandingImage } from "./components/LandingImage";
-import Activities from "./components/Activities";
-import Partners from "./components/Partners";
+import Activities from "./components/Activities/Activities";
+import Partners from "./components/Partners/Partners";
 import { Box } from "@mui/material";
 
 export default function Home() {
