+++ src/pages/home/components/Partners/constants.js
@@ -1,0 +1,20 @@
+import bombay from "assets/sponsors/bombay.png"
+import evo from "assets/sponsors/evo.png"
+import hellofresh from "assets/sponsors/hellofresh.png"
+import panela_lemon from "assets/sponsors/panela_lemon.jpg"
+import righteous from "assets/sponsors/righteous.png"
+import susgrainable from "assets/sponsors/susgrainable.png"
+import vnysa from "assets/sponsors/vnysa.png"
+import rainbow from "assets/rainbow.svg";
+
+const partners = [
+    { 'name': 'Bombay Masala', 'logo': bombay },
+    { 'name': 'Evo', 'logo': evo },
+    { 'name': 'Hello Fresh', 'logo': hellofresh },
+    { 'name': 'Panela Lemon', 'logo': panela_lemon },
+    { 'name': 'Righteous', 'logo': righteous },
+    { 'name': 'Susgrainable', 'logo': susgrainable },
+    { 'name': 'Vnysa', 'logo': vnysa }
+]
+
+export {partners, rainbow}
